Refactor, extracting shared logic and adding `readJSON` to TaskContext
