+++ apps/backend/src/shared/utils.ts
